Add files via upload
I found that when I attempted to utilize the sine, cosine and tangent functions in R, the results were not accurate and I was unable to use them in my project. For that reason I started from scratch and am in the process of creating trigonometry functions for use in Ultrasonics.